Hsearch (#1048)
* move hsearch into search.hybrid

* rename hsearch -> search.hybrid

* fake api to support original hsearch

* hsearch -> search.hybird in integration

* _hybrid.py -> _searcher.py
