+++ integrations/dense/test_dpr.py
@@ -62,7 +62,7 @@ class TestSearchIntegration(unittest.TestCase):
         output_file = 'test_run.dpr.nq-test.multi.bf.otf.bm25.trec'
         retrieval_file = 'test_run.dpr.nq-test.multi.bf.otf.bm25.json'
         self.temp_files.extend([output_file, retrieval_file])
-        cmd1 = f'python -m pyserini.hsearch dense  --index wikipedia-dpr-multi-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index wikipedia-dpr-multi-bf \
                                     --encoder facebook/dpr-question_encoder-multiset-base \
                              sparse --index wikipedia-dpr \
                              fusion --alpha 1.3 \
@@ -114,7 +114,7 @@ class TestSearchIntegration(unittest.TestCase):
         output_file = 'test_run.dpr.trivia-test.multi.bf.otf.bm25.trec'
         retrieval_file = 'test_run.dpr.trivia-test.multi.bf.otf.bm25.json'
         self.temp_files.extend([output_file, retrieval_file])
-        cmd1 = f'python -m pyserini.hsearch dense  --index wikipedia-dpr-multi-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index wikipedia-dpr-multi-bf \
                                     --encoder facebook/dpr-question_encoder-multiset-base \
                              sparse --index wikipedia-dpr \
                              fusion --alpha 0.95 \
@@ -166,7 +166,7 @@ class TestSearchIntegration(unittest.TestCase):
         output_file = 'test_run.dpr.wq-test.multi.bf.otf.bm25.trec'
         retrieval_file = 'test_run.dpr.wq-test.multi.bf.otf.bm25.json'
         self.temp_files.extend([output_file, retrieval_file])
-        cmd1 = f'python -m pyserini.hsearch dense  --index wikipedia-dpr-multi-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index wikipedia-dpr-multi-bf \
                                     --encoder facebook/dpr-question_encoder-multiset-base \
                              sparse --index wikipedia-dpr \
                              fusion --alpha 0.95 \
@@ -219,7 +219,7 @@ class TestSearchIntegration(unittest.TestCase):
         output_file = 'test_run.dpr.curated-test.multi.bf.otf.bm25.trec'
         retrieval_file = 'test_run.dpr.curated-test.multi.bf.otf.bm25.json'
         self.temp_files.extend([output_file, retrieval_file])
-        cmd1 = f'python -m pyserini.hsearch dense  --index wikipedia-dpr-multi-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index wikipedia-dpr-multi-bf \
                                     --encoder facebook/dpr-question_encoder-multiset-base \
                              sparse --index wikipedia-dpr \
                              fusion --alpha 1.05 \
@@ -272,7 +272,7 @@ class TestSearchIntegration(unittest.TestCase):
         output_file = 'test_run.dpr.squad-test.multi.bf.otf.bm25.trec'
         retrieval_file = 'test_run.dpr.squad-test.multi.bf.otf.bm25.json'
         self.temp_files.extend([output_file, retrieval_file])
-        cmd1 = f'python -m pyserini.hsearch dense  --index wikipedia-dpr-multi-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index wikipedia-dpr-multi-bf \
                                     --encoder facebook/dpr-question_encoder-multiset-base \
                              sparse --index wikipedia-dpr \
                              fusion --alpha 2.0 \
