+++ integrations/dense/test_tct_colbert-v2.py
@@ -90,7 +90,7 @@ class TestSearchIntegration(unittest.TestCase):
     def test_msmarco_passage_tct_colbert_v2_hnp_bf_bm25_hybrid_otf(self):
         output_file = 'test_run.msmarco-passage.tct_colbert-v2-hnp.bf-otf.bm25.tsv'
         self.temp_files.append(output_file)
-        cmd1 = f'python -m pyserini.hsearch dense  --index msmarco-passage-tct_colbert-v2-hnp-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index msmarco-passage-tct_colbert-v2-hnp-bf \
                                     --encoder castorini/tct_colbert-v2-hnp-msmarco \
                              sparse --index msmarco-passage \
                              fusion --alpha 0.06 \
@@ -108,7 +108,7 @@ class TestSearchIntegration(unittest.TestCase):
     def test_msmarco_passage_tct_colbert_v2_hnp_bf_d2q_hybrid_otf(self):
         output_file = 'test_run.msmarco-passage.tct_colbert-v2-hnp.bf-otf.doc2queryT5.tsv'
         self.temp_files.append(output_file)
-        cmd1 = f'python -m pyserini.hsearch dense  --index msmarco-passage-tct_colbert-v2-hnp-bf \
+        cmd1 = f'python -m pyserini.search.hybrid dense  --index msmarco-passage-tct_colbert-v2-hnp-bf \
                                     --encoder castorini/tct_colbert-v2-hnp-msmarco \
                              sparse --index msmarco-passage-expanded \
                              fusion --alpha 0.1 \
