+++ pyserini/demo/dpr.py
@@ -20,7 +20,7 @@ import random
 
 from pyserini.search.lucene import LuceneSearcher
 from pyserini.search.faiss import FaissSearcher, DprQueryEncoder
-from pyserini.hsearch import HybridSearcher
+from pyserini.search.hybrid import HybridSearcher
 from pyserini import search
