+++ pyserini/search/hybrid/__init__.py
@@ -14,6 +14,6 @@
 # limitations under the License.
 #
 
-from pyserini.search.hybrid import HybridSearcher
+from ._searcher import HybridSearcher
 
 __all__ = ['HybridSearcher']
